changed naming to ChasRooms instead to lessen on errors
