+++ ChasRooms.Server/Extensions.cs
