+++ Piglin-Brutes-Hackathon.AppHost/AppHost.cs
@@ -2,7 +2,7 @@ var builder = DistributedApplication.CreateBuilder(args);
 
 var cache = builder.AddRedis("cache");
 
-var server = builder.AddProject<Projects.Piglin_Brutes_Hackathon_Server>("server")
+var server = builder.AddProject<Projects.ChasRooms_Server>("server")
     .WithReference(cache)
     .WaitFor(cache)
     .WithHttpHealthCheck("/health")
