--- Piglin-Brutes-Hackathon.Server/Domain/Entities/Room.cs
@@ -1,12 +1,0 @@
-﻿namespace Piglin-Brutes-Hackathon.Server.Domain.Entities
-{
-    public class Room
-{
-    public int Id { get; set; }
-    public string RoomName { get; set; } = null!;
-    public int Capacity { get; set; }
-    public string features { get; set; } = null!;
-
-    ICollection<Booking> Bookings { get; set; } = new List<Booking>();
-}
-}
