--- Piglin-Brutes-Hackathon.Server/Domain/Entities/UserBooking.cs
@@ -1,11 +1,0 @@
-﻿namespace Piglin-Brutes-Hackathon.Server.Domain.Entities
-{
-    public class UserBooking
-{
-    public Guid BookingId { get; set; }
-    public Booking Booking { get; set; } = null!;
-
-    public User User { get; set; } = null!;
-    public string UserId { get; set; } = null!;
-}
-}
